(PC-20011)[API] refactor: split file routes/public/collective/serialization/offers
