+++ api/src/pcapi/routes/public/collective/endpoints/domains.py
@@ -2,6 +2,7 @@ from typing import cast
 
 from pcapi.core.educational import repository as educational_repository
 from pcapi.routes.public import blueprints
+from pcapi.routes.public.collective.serialization import domains as domains_serialization
 from pcapi.routes.public.collective.serialization import offers as offers_serialization
 from pcapi.routes.serialization import BaseModel
 from pcapi.serialization.decorator import spectree_serialize
@@ -18,7 +19,7 @@ from pcapi.validation.routes.users_authentifications import api_key_required
         **(
             {
                 "HTTP_200": (
-                    offers_serialization.CollectiveOffersListDomainsResponseModel,
+                    domains_serialization.CollectiveOffersListDomainsResponseModel,
                     "La liste des domaines d'éducation.",
                 ),
                 "HTTP_401": (
@@ -31,14 +32,14 @@ from pcapi.validation.routes.users_authentifications import api_key_required
 )
 @api_key_required
 @cached_view(prefix="pro_public_api_v2")
-def list_educational_domains() -> offers_serialization.CollectiveOffersListDomainsResponseModel:
+def list_educational_domains() -> domains_serialization.CollectiveOffersListDomainsResponseModel:
     # in French, to be used by Swagger for the API documentation
     """Récupération de la liste des domaines d'éducation pouvant être associés aux offres collectives."""
     educational_domains = educational_repository.get_all_educational_domains_ordered_by_name()
 
-    return offers_serialization.CollectiveOffersListDomainsResponseModel(
+    return domains_serialization.CollectiveOffersListDomainsResponseModel(
         __root__=[
-            offers_serialization.CollectiveOffersDomainResponseModel.from_orm(educational_domain)
+            domains_serialization.CollectiveOffersDomainResponseModel.from_orm(educational_domain)
             for educational_domain in educational_domains
         ]
     )
