+++ api/src/pcapi/routes/public/collective/serialization/students_levels.py
@@ -1,0 +1,10 @@
+from pcapi.routes.serialization import BaseModel
+
+
+class CollectiveOffersStudentLevelResponseModel(BaseModel):
+    id: str
+    name: str
+
+
+class CollectiveOffersListStudentLevelsResponseModel(BaseModel):
+    __root__: list[CollectiveOffersStudentLevelResponseModel]
